Disable default-features for git2
The git2 default-features all have to do with ssh or ssh-keys. Those
require OpenSSL, which seems to be a bit tricky to get to compile
properly when cross-compiling on CI.

Also, since this project does not need those parts of the git2
dependency, we should be able to safely disable those features.

Co-authored-by: Maria Hansson <mariahansson92@hotmail.com>