Stop printing on stderr when adding first coauthor
Before this change, when running `git add-coauthor` without an existing
`~/.git-coauthors` file, the program would print an error message on
stderr that the file did not exist.

It would then create the file with the newly added co-author in it.

That message seemed unnecessary and so it has now been removed.

Co-authored-by: Maria Hansson <mariahansson92@hotmail.com>